requirements.txt
slack-sdk
slack-bolt
watchfiles
pydantic
httpx
aiohttp
fastapi
google-auth
requests
google-cloud-storage
python-dotenv
google-cloud-aiplatform